silence compiler warnings about error format strings in Octave 5
* cl2bp.cc (Fcl2bp): Use "%s" to format arbitrary string to silence compiler
warning about format security with non literal strings.
* medfilt1.cc (Fmedfilt1): Use "%ld" and static_cast<long> to format argument
of system-dependent type octave_idx_type to silence compiler warning about
format argument type.

+++ src/medfilt1.cc
@@ -360,7 +360,8 @@ to bring them up to size @var{n}.\n\
                      args(3).double_value ());
             //if (dim < 1 || dim > signal.dims ().length ())
             if (dim < 1)
-              error ("medfilt1: DIM must be positive, not %d", dim);
+              error ("medfilt1: DIM must be positive, not %ld",
+                     static_cast<long> (dim));
           }
         else
           error ("medfilt1: Invalid type for DIM: %s",
